Created CourseAssessmentMarks class

+++ Course.cpp
@@ -1,6 +1,7 @@
 #include "Course.h"
+#include "Department.h"
 
-Course::Course(Department* department, int code, std::string name, std::string description, int credits, int fees) {
+Course::Course(Department* department, int code, std::string name, std::string description, int credits, double fees) {
 	_department = department;
 	_code = code;
 	_name = name;
